feat: Data version control
In an effort to make the data more accessable and reproducible I've
adopted [Data Version Control](https://dvc.org/) or dvc.

Since I was moving to versioned data I also took this opporunity to
change the format for all of my processed data to be in an accessable,
standarized format; the goal is that even if the current source code is
unusable, the data will hopefully still be retrievable, because it is
represented in a very standard format that many programs now and into
the feature should be able to read.